move out shared functionality
